fully completed
Everything is functional.
improvements that would be nice:
On certain frames, the entire thing is just dots, I'm not too sure why it is. It could be that it is shifting the donut far away and as a result al points are at an infinite distance, however, surely some would be closer than others due to the spherical nature.
Another would be to clean up the maths, I manually did all of the dot products for the sake of learning it, however, this can be cleaned up by using the "dot." funcition in maths.